Fixing "warning: comparison between signed and unsigned integer expressions" (hopefully).
git-svn-id: http://svn.seqan.de/seqan/trunk@6345 e6417c60-b987-48fd-844e-b20f0fcc1017

+++ seqan/projects/library/seqan/seeds/memoryManager_base.h
@@ -329,7 +329,7 @@ obtainID(MemoryManager<TValue,Block<SPACE>,TFree > &manager)
 	}else{
 		int i = 0;
 		long tmp = (TValue*)manager.pEmptySpace - (TValue*)manager.blocks[i];
-		while ((tmp >=SPACE)||( tmp < 0)){
+		while ((tmp >= static_cast<long>(SPACE))||( tmp < 0)){
 			++i;
 			tmp = (TValue*)manager.pEmptySpace - (TValue*)manager.blocks[i];
 		}
@@ -346,7 +346,7 @@ _getPosition(MemoryManager<TValue,Block<SPACE>, TFree > const &manager,
 	SEQAN_CHECKPOINT
 	int i = 0;
 	long tmp = (TValue*)pointer - (TValue*)manager.blocks[i];
-	while ((tmp >=SPACE)||( tmp < 0)){
+	while ((tmp >= static_cast<long>(SPACE))||( tmp < 0)){
 		++i;
 		tmp = (TValue*)pointer - (TValue*)manager.blocks[i];
 	}
